feat(pci-instance): add new routes
